+++ UriModifier.php
@@ -13,73 +13,18 @@ declare(strict_types=1);
 
 namespace League\Uri;
 
-use BadMethodCallException;
-use League\Uri\Components\DataPath;
-use League\Uri\Components\Domain;
-use League\Uri\Components\HierarchicalPath;
-use League\Uri\Components\Host;
-use League\Uri\Components\Path;
-use League\Uri\Components\Query;
-use League\Uri\Contracts\PathInterface;
 use League\Uri\Contracts\UriInterface;
 use League\Uri\Exceptions\SyntaxError;
 use Psr\Http\Message\UriInterface as Psr7UriInterface;
-use ReflectionClass;
-use ReflectionMethod;
 use Stringable;
-use function ltrim;
-use function rtrim;
 
+/**
+ * @deprecated since version 7.0.0
+ * @codeCoverageIgnore
+ * @see Modifier
+ */
 class UriModifier
 {
-    public static function from(Stringable|string $uri): self
-    {
-        return new static(static::filterUri($uri));
-    }
-
-    private static function modifiers(string $methodName): ?ReflectionMethod
-    {
-        static $methods = null;
-
-        if (null === $methods) {
-            foreach ((new ReflectionClass(static::class))->getMethods(ReflectionMethod::IS_STATIC) as $method) {
-                if ($method->isPublic() && 'from' !== $method->getName()) {
-                    $methods[$method->getName()] = $method;
-                }
-            }
-        }
-
-        return $methods[$methodName] ?? null;
-    }
-
-    final protected function __construct(protected readonly Psr7UriInterface|UriInterface $uri)
-    {
-    }
-
-    /**
-     * @param string $method       the name of the static modifier method
-     * @param mixed  ...$arguments arguments of the static modifier method
-     *
-     * @throws BadMethodCallException if the method is not defined
-     */
-    final public function pipe(string $method, mixed ...$arguments): self
-    {
-        $modifier = self::modifiers($method);
-        if (null === $modifier) {
-            throw new BadMethodCallException('The modifier method '.self::class.'::'.$method.' does not exist.');
-        }
-
-        /** @var Psr7UriInterface|UriInterface $uri */
-        $uri = $modifier->invoke(null, $this->uri, ...$arguments);
-
-        return new self($uri);
-    }
-
-    final public function get(): Psr7UriInterface|UriInterface
-    {
-        return $this->uri;
-    }
-
     /*********************************
      * Query resolution methods
      *********************************/
@@ -91,11 +36,7 @@ class UriModifier
         Stringable|string $uri,
         Stringable|string|null $query
     ): Psr7UriInterface|UriInterface {
-        $uri = static::filterUri($uri);
-
-        return $uri->withQuery(
-            static::normalizeComponent(Query::fromUri($uri)->append($query)->value(), $uri)
-        );
+        return Modifier::from($uri)->appendQuery($query)->get();
     }
 
     /**
@@ -105,11 +46,7 @@ class UriModifier
         Stringable|string $uri,
         Stringable|string|null $query
     ): Psr7UriInterface|UriInterface {
-        $uri = static::filterUri($uri);
-
-        return $uri->withQuery(
-            static::normalizeComponent(Query::fromUri($uri)->merge($query)->value(), $uri)
-        );
+        return Modifier::from($uri)->mergeQuery($query)->get();
     }
 
     /**
@@ -117,11 +54,7 @@ class UriModifier
      */
     public static function removePairs(Stringable|string $uri, string ...$keys): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-
-        return $uri->withQuery(
-            static::normalizeComponent(Query::fromUri($uri)->withoutPair(...$keys)->value(), $uri)
-        );
+        return Modifier::from($uri)->removePairs(...$keys)->get();
     }
 
     /**
@@ -132,11 +65,7 @@ class UriModifier
      */
     public static function removeEmptyPairs(Stringable|string $uri): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-
-        return $uri->withQuery(
-            static::normalizeComponent(Query::fromUri($uri)->withoutEmptyPairs()->value(), $uri)
-        );
+        return Modifier::from($uri)->removeEmptyPairs()->get();
     }
 
     /**
@@ -144,11 +73,7 @@ class UriModifier
      */
     public static function removeParams(Stringable|string $uri, string ...$keys): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-
-        return $uri->withQuery(
-            static::normalizeComponent(Query::fromUri($uri)->withoutParameters(...$keys)->value(), $uri)
-        );
+        return Modifier::from($uri)->removeParams(...$keys)->get();
     }
 
     /**
@@ -156,9 +81,7 @@ class UriModifier
      */
     public static function sortQuery(Stringable|string $uri): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-
-        return $uri->withQuery(static::normalizeComponent(Query::fromUri($uri)->sort()->value(), $uri));
+        return Modifier::from($uri)->sortQuery()->get();
     }
 
     /*********************************
@@ -170,13 +93,7 @@ class UriModifier
      */
     public static function addRootLabel(Stringable|string $uri): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-        $host = Domain::fromUri($uri)->withRootLabel()->value();
-        if (null === $host || $host === $uri->getHost() || !str_ends_with($host, '.')) {
-            return $uri;
-        }
-
-        return $uri->withHost($uri->getHost().'.');
+        return Modifier::from($uri)->addRootLabel()->get();
     }
 
     /**
@@ -186,16 +103,7 @@ class UriModifier
      */
     public static function appendLabel(Stringable|string $uri, Stringable|string|null $label): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-        $host = Host::fromUri($uri);
-        $label = Host::new($label);
-
-        return match (true) {
-            null === $label->value() => $uri,
-            $host->isDomain() => $uri->withHost(static::normalizeComponent(Domain::new($host)->append($label)->value(), $uri)),
-            $host->isIpv4() => $uri->withHost($host->value().'.'.ltrim($label->value(), '.')),
-            default => throw new SyntaxError('The URI host '.$host->toString().' can not be appended.'),
-        };
+        return Modifier::from($uri)->appendLabel($label)->get();
     }
 
     /**
@@ -203,9 +111,7 @@ class UriModifier
      */
     public static function hostToAscii(Stringable|string $uri): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-
-        return $uri->withHost(static::normalizeComponent(Host::fromUri($uri)->value(), $uri));
+        return Modifier::from($uri)->hostToAscii()->get();
     }
 
     /**
@@ -213,9 +119,7 @@ class UriModifier
      */
     public static function hostToUnicode(Stringable|string $uri): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-
-        return $uri->withHost(static::normalizeComponent(Host::fromUri($uri)->toUnicode(), $uri));
+        return Modifier::from($uri)->hostToUnicode()->get();
     }
 
     /**
@@ -225,16 +129,7 @@ class UriModifier
      */
     public static function prependLabel(Stringable|string $uri, Stringable|string|null $label): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-        $host = Host::fromUri($uri);
-        $label = Host::new($label);
-
-        return match (true) {
-            null === $label->value() => $uri,
-            $host->isDomain() => $uri->withHost(static::normalizeComponent(Domain::new($host)->prepend($label)->value(), $uri)),
-            $host->isIpv4() => $uri->withHost(rtrim($label->value(), '.').'.'.$host->value()),
-            default => throw new SyntaxError('The URI host '.$host->toString().' can not be prepended.'),
-        };
+        return Modifier::from($uri)->prependLabel($label)->get();
     }
 
     /**
@@ -242,11 +137,7 @@ class UriModifier
      */
     public static function removeLabels(Stringable|string $uri, int ...$keys): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-
-        return $uri->withHost(
-            static::normalizeComponent(Domain::fromUri($uri)->withoutLabel(...$keys)->value(), $uri)
-        );
+        return Modifier::from($uri)->removeLabels(...$keys)->get();
     }
 
     /**
@@ -254,16 +145,7 @@ class UriModifier
      */
     public static function removeRootLabel(Stringable|string $uri): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-        $currentHost = $uri->getHost();
-        if (null === $currentHost || '' === $currentHost || !str_ends_with($currentHost, '.')) {
-            return $uri;
-        }
-
-        /** @var string $host */
-        $host = Domain::new($currentHost)->value();
-
-        return $uri->withHost($host);
+        return Modifier::from($uri)->removeRootLabel()->get();
     }
 
     /**
@@ -271,11 +153,7 @@ class UriModifier
      */
     public static function removeZoneId(Stringable|string $uri): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-
-        return $uri->withHost(
-            static::normalizeComponent(Host::fromUri($uri)->withoutZoneIdentifier()->value(), $uri)
-        );
+        return Modifier::from($uri)->removeZoneId()->get();
     }
 
     /**
@@ -286,11 +164,7 @@ class UriModifier
         int $offset,
         Stringable|string|null $label
     ): Psr7UriInterface|UriInterface {
-        $uri = static::filterUri($uri);
-
-        return $uri->withHost(
-            static::normalizeComponent(Domain::fromUri($uri)->withLabel($offset, $label)->value(), $uri)
-        );
+        return Modifier::from($uri)->replaceLabel($offset, $label)->get();
     }
 
     /*********************************
@@ -302,16 +176,7 @@ class UriModifier
      */
     public static function addBasePath(Stringable|string $uri, Stringable|string $path): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-        /** @var HierarchicalPath $path */
-        $path = HierarchicalPath::new($path)->withLeadingSlash();
-        /** @var HierarchicalPath $currentPath */
-        $currentPath = HierarchicalPath::fromUri($uri)->withLeadingSlash();
-
-        return match (true) {
-            !str_starts_with($currentPath->toString(), $path->toString()) => $uri->withPath($path->append($currentPath)->toString()),
-            default => static::normalizePath($uri, $currentPath),
-        };
+        return Modifier::from($uri)->addBasePath($path)->get();
     }
 
     /**
@@ -319,9 +184,7 @@ class UriModifier
      */
     public static function addLeadingSlash(Stringable|string $uri): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-
-        return $uri->withPath(Path::fromUri($uri)->withLeadingSlash()->toString());
+        return Modifier::from($uri)->addLeadingSlash()->get();
     }
 
     /**
@@ -329,21 +192,15 @@ class UriModifier
      */
     public static function addTrailingSlash(Stringable|string $uri): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-
-        return $uri->withPath(Path::fromUri($uri)->withTrailingSlash()->toString());
+        return Modifier::from($uri)->addTrailingSlash()->get();
     }
 
     /**
      * Append a new segment or a new path to the URI path.
      */
-    public static function appendSegment(
-        Stringable|string $uri,
-        Stringable|string $segment
-    ): Psr7UriInterface|UriInterface {
-        $uri = static::filterUri($uri);
-
-        return static::normalizePath($uri, HierarchicalPath::fromUri($uri)->append($segment));
+    public static function appendSegment(Stringable|string $uri, Stringable|string $segment): Psr7UriInterface|UriInterface
+    {
+        return Modifier::from($uri)->appendSegment($segment)->get();
     }
 
     /**
@@ -351,9 +208,7 @@ class UriModifier
      */
     public static function dataPathToAscii(Stringable|string $uri): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-
-        return $uri->withPath(DataPath::fromUri($uri)->toAscii()->toString());
+        return Modifier::from($uri)->dataPathToAscii()->get();
     }
 
     /**
@@ -361,9 +216,7 @@ class UriModifier
      */
     public static function dataPathToBinary(Stringable|string $uri): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-
-        return $uri->withPath(DataPath::fromUri($uri)->toBinary()->toString());
+        return Modifier::from($uri)->dataPathToBinary()->get();
     }
 
     /**
@@ -373,9 +226,7 @@ class UriModifier
         Stringable|string $uri,
         Stringable|string $segment
     ): Psr7UriInterface|UriInterface {
-        $uri = static::filterUri($uri);
-
-        return static::normalizePath($uri, HierarchicalPath::fromUri($uri)->prepend($segment));
+        return Modifier::from($uri)->prependSegment($segment)->get();
     }
 
     /**
@@ -385,17 +236,7 @@ class UriModifier
         Stringable|string $uri,
         Stringable|string $path
     ): Psr7UriInterface|UriInterface {
-        $uri = static::filterUri($uri);
-        $basePath = HierarchicalPath::new($path)->withLeadingSlash()->toString();
-        $currentPath = HierarchicalPath::fromUri($uri)->withLeadingSlash()->toString();
-        $newPath = substr($currentPath, strlen($basePath));
-
-        return match (true) {
-            '/' === $basePath,
-            !str_starts_with($currentPath, $basePath),
-            !str_starts_with($newPath, '/') => $uri,
-            default => $uri->withPath($newPath),
-        };
+        return Modifier::from($uri)->removeBasePath($path)->get();
     }
 
     /**
@@ -403,9 +244,7 @@ class UriModifier
      */
     public static function removeDotSegments(Stringable|string $uri): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-
-        return $uri->withPath(Path::fromUri($uri)->withoutDotSegments()->toString());
+        return Modifier::from($uri)->removeDotSegments()->get();
     }
 
     /**
@@ -413,9 +252,7 @@ class UriModifier
      */
     public static function removeEmptySegments(Stringable|string $uri): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-
-        return $uri->withPath(HierarchicalPath::fromUri($uri)->withoutEmptySegments()->toString());
+        return Modifier::from($uri)->removeEmptySegments()->get();
     }
 
     /**
@@ -423,9 +260,8 @@ class UriModifier
      */
     public static function removeLeadingSlash(Stringable|string $uri): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
+        return Modifier::from($uri)->removeLeadingSlash()->get();
 
-        return static::normalizePath($uri, Path::fromUri($uri)->withoutLeadingSlash());
     }
 
     /**
@@ -433,9 +269,7 @@ class UriModifier
      */
     public static function removeTrailingSlash(Stringable|string $uri): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-
-        return $uri->withPath(Path::fromUri($uri)->withoutTrailingSlash()->toString());
+        return Modifier::from($uri)->removeTrailingSlash()->get();
     }
 
     /**
@@ -443,9 +277,7 @@ class UriModifier
      */
     public static function removeSegments(Stringable|string $uri, int ...$keys): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-
-        return $uri->withPath(HierarchicalPath::fromUri($uri)->withoutSegment(...$keys)->toString());
+        return Modifier::from($uri)->removeSegments(...$keys)->get();
     }
 
     /**
@@ -453,9 +285,7 @@ class UriModifier
      */
     public static function replaceBasename(Stringable|string $uri, Stringable|string $basename): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-
-        return static::normalizePath($uri, HierarchicalPath::fromUri($uri)->withBasename($basename));
+        return Modifier::from($uri)->replaceBasename($basename)->get();
     }
 
     /**
@@ -463,9 +293,7 @@ class UriModifier
      */
     public static function replaceDataUriParameters(Stringable|string $uri, Stringable|string $parameters): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-
-        return $uri->withPath(DataPath::fromUri($uri)->withParameters($parameters)->toString());
+        return Modifier::from($uri)->replaceDataUriParameters($parameters)->get();
     }
 
     /**
@@ -473,9 +301,7 @@ class UriModifier
      */
     public static function replaceDirname(Stringable|string $uri, Stringable|string $dirname): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-
-        return static::normalizePath($uri, HierarchicalPath::fromUri($uri)->withDirname($dirname));
+        return Modifier::from($uri)->replaceDirname($dirname)->get();
     }
 
     /**
@@ -483,9 +309,7 @@ class UriModifier
      */
     public static function replaceExtension(Stringable|string $uri, Stringable|string $extension): Psr7UriInterface|UriInterface
     {
-        $uri = static::filterUri($uri);
-
-        return $uri->withPath(HierarchicalPath::fromUri($uri)->withExtension($extension)->toString());
+        return Modifier::from($uri)->replaceExtension($extension)->get();
     }
 
     /**
@@ -496,53 +320,6 @@ class UriModifier
         int $offset,
         Stringable|string $segment
     ): Psr7UriInterface|UriInterface {
-        $uri = static::filterUri($uri);
-
-        return $uri->withPath(HierarchicalPath::fromUri($uri)->withSegment($offset, $segment)->toString());
-    }
-
-    /**
-     * Input URI normalization to allow Stringable and string URI.
-     */
-    final protected static function filterUri(Stringable|string $uri): Psr7UriInterface|UriInterface
-    {
-        return match (true) {
-            $uri instanceof BaseUri => $uri->uri(),
-            $uri instanceof Psr7UriInterface, $uri instanceof UriInterface => $uri,
-            default => Uri::new($uri),
-        };
-    }
-
-    /**
-     * Normalize a URI path.
-     *
-     * Make sure the path always has a leading slash if an authority is present
-     * and the path is not the empty string.
-     */
-    final protected static function normalizePath(Psr7UriInterface|UriInterface $uri, PathInterface $path): Psr7UriInterface|UriInterface
-    {
-        $pathString = $path->toString();
-        $authority = $uri->getAuthority();
-
-        return match (true) {
-            '' === $pathString,
-            '/' === $pathString[0],
-            null === $authority,
-            '' === $authority => $uri->withPath($pathString),
-            default => $uri->withPath('/'.$pathString),
-        };
-    }
-
-    /**
-     * Normalize the URI component value depending on the subject interface.
-     *
-     * null value MUST be converted to the emptu string if a Psr7 UriInterface is being manipulated.
-     */
-    final protected static function normalizeComponent(?string $component, Psr7UriInterface|UriInterface $uri): ?string
-    {
-        return match (true) {
-            $uri instanceof Psr7UriInterface => (string) $component,
-            default => $component,
-        };
+        return Modifier::from($uri)->replaceSegment($offset, $segment)->get();
     }
 }