+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All Notable changes to `League\Uri\Components` will be documented in this file
 - `League\Uri\Components\Query::hasParameter` accessing method
 - `League\Uri\IPv4Calculators\MissingIPv4Calculator` exception
 - `League\Uri\Components\OffsetOutOfBounds` exception added
+- `League\Uri\Modifier` to replace deprecated `League\Uri\Modifier`
 
 ### Fixed
 
@@ -44,6 +45,7 @@ All Notable changes to `League\Uri\Components` will be documented in this file
 - `Query::params` is replaced by `Query::parameters` and `Query::parameter`
 - `HierarchicalPath::segments` use `HierarchicalPath::getIterator` instead
 - `Domain::labels` use `Domain::getIterator` instead
+- `League\Uri\Modifier`
 
 ### Removed